+++ R/cusum.R
@@ -8,6 +8,7 @@
 #' @param failure_probability Double. Baseline failure probability
 #' @param patient_outcomes Integer. Vector of binary patient outcomes (0,1)
 #' @param limit Double. Control limit for signalling performance change
+#' @param weights Double. Optional vector of weights, if empty, standard CUSUM weights are calculated with weights_t
 #' @param odds_multiplier Double. Odds multiplier of adverse event under the alternative hypothesis (<1 looks for decreases)
 #' @param reset Logical. Reset the CUSUM after a signal to 0; defaults to TRUE
 #' @examples
@@ -37,7 +38,7 @@
 #'   patient_outcomes,
 #'   limit = 2.96
 #' )
-cusum <- function(failure_probability, patient_outcomes, limit, odds_multiplier = 2, reset = TRUE) {
+cusum <- function(failure_probability, patient_outcomes, limit, weights = NA, odds_multiplier = 2, reset = TRUE) {
 
   ## Check user input ####
   assert_numeric(failure_probability, lower = 0, upper = 1, finite = TRUE, any.missing = FALSE, len = 1)
@@ -51,6 +52,13 @@ cusum <- function(failure_probability, patient_outcomes, limit, odds_multiplier 
 
   assert_numeric(limit, finite = TRUE, any.missing = FALSE, len = 1)
 
+  if (!is.na(weights)){
+    assert_numeric(weights, lower = 0, upper = 1, finite = TRUE, any.missing = FALSE, min.len = 1)
+    if (length(weights) != length(patient_outcomes)) {
+      stop("Length weights and patient outcomes of unequal size.")
+    }
+  }
+  
   assert_numeric(odds_multiplier, lower = 0, finite = TRUE, any.missing = FALSE, len = 1)
   if (odds_multiplier < 1) {
     message("CUSUM is set to detect process improvements (odds_multiplier < 1). ")
@@ -69,19 +77,15 @@ cusum <- function(failure_probability, patient_outcomes, limit, odds_multiplier 
   ## Calculate CUSUM Chart ####
   npat <- length(patient_outcomes)
 
-  p.0 <- failure_probability
-  o.0 <- p.0 / (1 - p.0)
-  o.1 <- odds_multiplier * o.0
-  p.1 <- o.1 / (1 + o.1)
-  ca <- p.1
-
   ct <- 0
   cs <- matrix(0, nrow = npat, ncol = 5)
 
-  wf <- log(ca / failure_probability) # CS weight for failure
-  ws <- log((1 - ca) / (1 - failure_probability)) # CS weight for success
+  if (is.na(weights)){
+    w <-  weights_t(patient_outcomes,
+                    probability_ae = failure_probability,
+                    odds_multiplier)
+  }
 
-  w <- ifelse(patient_outcomes == 1, wf, ws) # weights based on outcome
 
   for (ii in 1:npat) {