+++ R/racusum.R
@@ -8,6 +8,7 @@
 #' @param patient_risks Double. Vector of patient risk scores (individual risk of adverse event)
 #' @param patient_outcomes Integer. Vector of binary patient outcomes (0,1)
 #' @param limit Double. Control limit for signalling performance change
+#' @param weights Double. Optional vector of weights, if empty, standard CUSUM weights are calculated with weights_t
 #' @param odds_multiplier Double. Odds multiplier of adverse event under the alternative hypothesis (<1 looks for decreases)
 #' @param reset Logical. Reset the CUSUM after a signal to 0; defaults to TRUE
 #' @param limit_method "constant" or "dynamic"
@@ -50,13 +51,20 @@
 #'   patient_outcomes,
 #'   limit = 2.96
 #' )
-racusum <- function(patient_risks, patient_outcomes, limit, odds_multiplier = 2, reset = TRUE, limit_method = c("constant", "dynamic")) {
+racusum <- function(patient_risks, patient_outcomes, limit, weights = NA, odds_multiplier = 2, reset = TRUE, limit_method = c("constant", "dynamic")) {
   npat <- length(patient_risks)
 
   ## Check user input ####
   assert_numeric(patient_risks, lower = 0, upper = 1, finite = TRUE, any.missing = FALSE, min.len = 1)
   if (length(patient_risks) != length(patient_outcomes)) {
     stop("Length patient_risks and patient_outcomes of unequal size.")
+  }
+  
+  if (!is.na(weights)){
+    assert_numeric(weights, lower = 0, upper = 1, finite = TRUE, any.missing = FALSE, min.len = 1)
+    if (length(weights) != length(patient_outcomes)) {
+      stop("Length weights and patient outcomes of unequal size.")
+    }
   }
 
   patient_outcomes <- as.integer(patient_outcomes)
@@ -90,17 +98,24 @@ racusum <- function(patient_risks, patient_outcomes, limit, odds_multiplier = 2,
 
   ## Calculate RA-CUSUM Chart ####
   npat <- length(patient_risks)
-
-  p <- patient_risks
-
+  
+  if (is.na(weights)){
+    w <- weights_t(patient_outcomes,
+                   probability_ae = patient_risks,
+                   odds_multiplier)
+  }
+  
   ct <- 0 # initial CUSUM value
+ 
   cs <- matrix(0, nrow = npat, ncol = 5) # storage matrix for alarms
-
-  # CUSUM weights
-  ws <- log((1 - p + 1 * p) / (1 - p + odds_multiplier * p))
-  wf <- log(((1 - p + 1 * p) * odds_multiplier) / ((1 - p + odds_multiplier * p) * 1))
-
-  w <- ifelse(patient_outcomes == 1, wf, ws) # weights based on outcome
+  # p <- patient_risks
+  # 
+   # 
+  # # CUSUM weights
+  # ws <- log((1 - p + 1 * p) / (1 - p + odds_multiplier * p))
+  # wf <- log(((1 - p + 1 * p) * odds_multiplier) / ((1 - p + odds_multiplier * p) * 1))
+  # 
+  # w <- ifelse(patient_outcomes == 1, wf, ws) # weights based on outcome
 
   for (ii in 1:npat) {
     if (odds_multiplier > 1) {
@@ -127,7 +142,7 @@ racusum <- function(patient_risks, patient_outcomes, limit, odds_multiplier = 2,
       }
     }
     cs[ii, 1] <- ii # store patient id
-    cs[ii, 2] <- p[ii] # store patient risk
+    cs[ii, 2] <- patient_risks[ii] # store patient risk
     cs[ii, 3] <- ct # store CUSUM value
   }